add left feature module to home page
